Virtual WebLab 10/19/2015
Virtual Lab

Build a 'sign up' web form.  You'll need to collect some information.
-First Name
-Last Name
-Password
-Title
-Favorite Beatle

In the form build select menus for the Title and Favorite Beatle
sections.

Titles:
   Mrs.
   Ms.
   Mr.
   Undetermined

Beatles:
   George
   Ringo
   Paul
   John

After the form submits gather the information and display it on the
page.

Also, using the information from the form, build a new username for
this person.

First two letters of the firstname +
Last two letters of the lastname +
Index of the selected Beatle (0 or 1 or 2 or 3)+
first two letters and last letter of the selected Beatles name.